FRONTEND-#47 small fixes
