+++ frontend/src/App.css
@@ -42,14 +42,15 @@
 }
 
 .course-card {
+  color: black;
   width: 300px;
   margin: 7px  7px 16px 5px;
   border-radius: 15px;
-  box-shadow: 0px 0px 2px 0px rgba(0, 0, 0, 0.5);
+  box-shadow: 0 0 2px 0 rgba(0, 0, 0, 0.5);
   cursor: pointer;
   justify-content: space-between;
 }
 
 .course-card:hover {
-  background: #f3f3f3
+  background: #f6f6f6
 }
