+++ frontend/src/App.tsx
@@ -1,42 +1,33 @@
 import './App.css'
 import CourseCard from "./components/CourseCard.tsx";
-import {useState} from "react";
+import Search from "./components/Search.tsx";
 
 function App() {
-  const [ search, setSearch ] = useState("");
-
-  const handleSearch = () => {
-      console.log(search);
-  }
+  
+  const courses = [
+    {rating: 4.5, semester: "112-1", title: "Probability and Statistics", instructor: "Lebron James"},
+    {rating: 4.5, semester: "112-1", title: "Probability and Statistics", instructor: "Lebron James"},
+    {rating: 4.5, semester: "112-1", title: "Probability and Statistics", instructor: "Lebron James"},
+  ]
 
   return (
       <>
           <div>
               <h1 className="font-bold text-center m-5 text-4xl">NTUEE Course Map</h1>
           </div>
-          <div className="text-center">
-              <select className="border-r-0 border p-2 rounded-l-lg h-10" name="semester" id="semester">
-                  <option value="112-1">112-1</option>
-                  <option value="111-2">111-2</option>
-                  <option value="111-1">111-1</option>
-                  <option value="110-2">110-2</option>
-                  <option value="110-1">110-1</option>
-              </select>
-              <input className="border mr-2 p-1.5 h-10 w-80"
-                     type="search"
-                     placeholder="search Course Map"
-                     value={search}
-                     onChange={(e) => setSearch(e.target.value)}
-              />
-              <button className="mx-1" onClick={handleSearch}>Search</button>
-          </div>
+          <Search/>
           <div>
               <div>
                   <h2 className="font-bold text-2xl mt-5">Popular</h2>
                   <div className="flex justify-center">
-                      <CourseCard/>
-                      <CourseCard/>
-                      <CourseCard/>
+                      {courses.map((course) => (
+                        <CourseCard
+                            title={course.title}
+                            instructor={course.instructor}
+                            rating={course.rating}
+                            semester={course.semester}
+                        />
+                      ))}
                   </div>
               </div>
               <div>
